requirements.txt
# RAG知识库系统依赖

# 核心框架
langchain>=0.1.0
langchain-community>=0.0.20

# 向量数据库
chromadb>=0.4.22

# Embedding模型
sentence-transformers>=2.2.2
torch>=2.0.0

# 文档处理
pyyaml>=6.0
markdown>=3.4.0
tiktoken>=0.5.0

# LLM集成（可选，根据需要选择）
openai>=1.0.0  # 如果使用OpenAI
# ollama>=0.1.0  # 如果使用Ollama本地模型

# API服务（可选）
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# 工具库
python-dotenv>=1.0.0  # 环境变量管理
tqdm>=4.66.0  # 进度条

# 检索增强（可选）
rank-bm25>=0.2.2  # BM25关键词检索
